pyproject.toml
[project]
name = "neuesvomtage"
version = "0.1.0"
description = "A news aggregator"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.5",
    "django-bootstrap5>=25.2",
    "django-cachalot>=2.8.0",
    "django-compressor>=4.5.1",
    "django-extensions>=4.1",
    "django-filter>=25.1",
    "django-libsass>=0.9",
    "favicon>=0.7.0",
    "feedparser>=6.0.11",
    "hypercorn>=0.17.3",
    "pillow>=11.3.0",
    "psycopg>=3.2.9",
    "pymemcache>=4.0.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "redis>=6.4.0",
    "requests>=2.32.4",
    "sorl-thumbnail>=12.11.0",
    "uritemplate>=4.2.0",
    "uvicorn>=0.35.0",
]
